Atahar Feature
I worked on the player mechanics for the game, focusing on movement and shooting functionality. The player's movement was controlled via scripting, which included both horizontal movement and jumping. To enhance the player experience, the camera was attached to the player in such a way that it follows their movement.

For the player’s death condition, the highest point they reached was used as the new frame of reference. If the player fell below this reference point, they would die. However, if the player moved above this point, it would be set as the new highest peak.

I also implemented a shooting mechanic for the player. The bullets would spawn directly above the player and were assigned a layer with a value of 8, including the layer name. Meanwhile, all enemies (represented as spheres or cubes) were assigned to layer 7 with the tag name "enemy," ensuring proper interaction between the player and enemies. This would also make sure that the players would lose instance if their layer number 7 would come into contact with layer number 8.